digestif: Fix startup by setting DIGESTIFPATH
Without this fix, digestif will just fail when run with the following
error:
    /nix/store/wg69xw20gnmyz171awp1gxg7y88fyibs-lua-5.3.6/bin/lua: ...5.3-digestif-0.2-1/share/lua/5.3/digestif/langserver.lua:341: Could not find data files at the following locations:
    - <homedir>/.local/share/digestif
    Set the environment variable DIGESTIFDATA to fix this.

Digestif ships with a bunch of "tag" files that contain information
about LaTeX commands. With the regular installer, these end up in
~/.local/share/digestif, but LuaRocks them under the data dir in the
rockspath. This patch adds sets the DIGESTIFPATH env var to point to the
correct location. Note that the user can generate tag files themselves
after installation, that why --set-default is used instead of --set